feat: 칼로리 계산 MET 지수로 변경

+++ backend/src/main/java/RunningMate/backend/domain/running/dto/RunningDTO.java
@@ -113,7 +113,7 @@ public class RunningDTO {
     public static class WhileRunningRequest {
         private Long recordId;
         private Duration runningTime;
-        private Long calories;
+        private Double calories;
         private Long distance;
 
     }
